Phase 4: split the game theater into useGameTheater + StageOverlay
`stageContent` is the app's second route: what is on the full-viewport
stage. Its state, the three effects that follow it (scroll lock,
`/play/<slug>` auto-open, the beforeunload guard), the five handlers that
open and close it, and 60 lines of overlay JSX were spread over 600 lines
of App.tsx.

- useGameTheater.ts  StageContent, stageContent + partyError, the three
  effects, and handlePlayGame / handleRemixGame /
  handleExitCatalogTheater / handleExitPartyTheater / handlePlayTogether
- StageOverlay.tsx   the party bar and GameTheater markup, plus the beta
  welcome splash that rides along with it, on explicit props

partyError moves into the hook rather than staying in App: handlePlay-
Together is its only writer and it exists to report that one failure.

App keeps the `stageOverlay` element and renders it from the same three
branches as before — the overlay must mount on the game-page and creator-
profile early returns too, or Play there scroll-locks the body behind
nothing. App.tsx 1112 -> 958.
